+++ changelog.md
@@ -21,6 +21,11 @@
 
 
 
+*	2020.9.17
+
+	+	sheetConstants.css: use CSS variables.
+
+
 *	2020.9.16
 
 	+	notation.ts: added parameter of trackList on notation output methods.